create msg directory tree for auth module
