+++ terra_sdk/core/auth/msgs/__init__.py
@@ -1,0 +1,9 @@
+from .msg_create_periodic_vesting_account import MsgCreatePeriodicVestingAccount
+from .msg_create_vesting_account import MsgCreateVestingAccount
+from .msg_donate_all_vesting_tokens import MsgDonateAllVestingTokens
+
+__all__ = [
+    "MsgCreatePeriodicVestingAccount",
+    "MsgCreateVestingAccount",
+    "MsgDonateAllVestingTokens"
+]
